requirements.txt
uuid
python-gitlab
requests
python-sonarqube-api